requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
drf-yasg==1.21.7
django-filter==24.3
celery==5.4.0
redis==5.0.7
gunicorn==21.2.0
dj-database-url==2.2.0
